Got things compiling on Windows in C++ mode. Lots of re-ordering of Clay UI element designated initializers. Some missing NEW_STRUCT(type) usages. Some missing type casts. Changed how CLAY_PACKED_ENUM macro works entirely. Added pntrConst to Str16 and used pntrConst in designated initializers in MakeStr8 and MakeStr16 macros so any pointer type will be accepted. Duplicated and rewrote OsDoOpenFileDialogBlocking and OsDoSaveFileDialog Windows implementations since the way we call member functions is different in C++.

+++ src/struct/struct_rich_string.h
@@ -639,14 +639,14 @@ PEXP RichStr DecodeStrToRichStr(Arena* arena, Str8 encodedString)
 							{
 								styleChangePartLength = (nextPipeIndex + 1) - bIndex;
 								styleChange.type = RichStrStyleChangeType_FontStyle;
-								if (valueBool) { styleChange.enableStyleFlags = styleFlag; FlagSet(state.enabledFlags, styleFlag); }
-								else { styleChange.disableStyleFlags = styleFlag; FlagUnset(state.enabledFlags, styleFlag); }
+								if (valueBool) { styleChange.enableStyleFlags = (u8)styleFlag; FlagSet(state.enabledFlags, styleFlag); }
+								else { styleChange.disableStyleFlags = (u8)styleFlag; FlagUnset(state.enabledFlags, styleFlag); }
 							}
 							else if (StrAnyCaseEquals(valuePart, StrLit("default")))
 							{
 								styleChangePartLength = (nextPipeIndex + 1) - bIndex;
 								styleChange.type = RichStrStyleChangeType_FontStyle;
-								styleChange.defaultStyleFlags = styleFlag;
+								styleChange.defaultStyleFlags = (u8)styleFlag;
 								FlagUnset(state.enabledFlags, styleFlag);
 							}
 						}
@@ -708,8 +708,8 @@ PEXP RichStr DecodeStrToRichStr(Arena* arena, Str8 encodedString)
 							
 							styleChangePartLength = (nextPipeIndex + 1) - bIndex;
 							styleChange.type = RichStrStyleChangeType_FontStyle;
-							if (IsFlagSet(state.enabledFlags, styleFlag)) { styleChange.defaultStyleFlags = styleFlag; FlagUnset(state.enabledFlags, styleFlag); }
-							else { styleChange.enableStyleFlags = styleFlag; FlagSet(state.enabledFlags, styleFlag); }
+							if (IsFlagSet(state.enabledFlags, styleFlag)) { styleChange.defaultStyleFlags = (u8)styleFlag; FlagUnset(state.enabledFlags, styleFlag); }
+							else { styleChange.enableStyleFlags = (u8)styleFlag; FlagSet(state.enabledFlags, styleFlag); }
 						}
 						else if (StrAnyCaseEquals(namePart, StrLit("color")) ||
 							StrAnyCaseEquals(namePart, StrLit("rgb")))
